support optional projectId on context; large refactor, componentize and upgrade deps
