--- src/index.test.tsx
@@ -1,7 +1,0 @@
-import { ExampleComponent } from '.'
-
-describe('ExampleComponent', () => {
-  it('is truthy', () => {
-    expect(ExampleComponent).toBeTruthy()
-  })
-})
