+++ tsconfig.json
@@ -4,7 +4,8 @@
     "module": "esnext",
     "lib": ["dom", "esnext"],
     "moduleResolution": "node",
-    "jsx": "react",
+    "jsx": "react-jsx",
+    "jsxFactory": null,
     "sourceMap": true,
     "declaration": true,
     "esModuleInterop": true,
